Add visualization widgets to catalog through publish.py
rubicon_ml.publish.py now optionally takes in a single viz object (ExperimentTable only at the moment) and publishs it as an entry in the output catalog. The arguments for the entry are all the arguments that are input into the viz object itself so it can be recreated later.