Add visualization widgets to catalog through publish.py (#450)
* Add visualization widgets to catalog through publish.py

rubicon_ml.publish.py now optionally takes in a single viz object (ExperimentTable only at the moment) and publishs it as an entry in the output catalog. The arguments for the entry are all the arguments that are input into the viz object itself so it can be recreated later.

* Visualization object input to test_publish.py

Added the ability to include a visualization object to this test which tests publish() from publish.py which at the moment has been updated to have the ability to publish the ExperimentsTable widget to catalogs.